refactor: extract workspace view from App
